Add files via upload
Hey guys added a bunch of visual elements to the information and stock we agreed on. For our meeting tomorrow I will have ready the use for the api. feel free to take a look before commiting to the main branch so any changes can be made before hand. Thanks